requirements.txt
streamlit
google-genai
pypdf
python-docx
feedparser
requests
beautifulsoup4
lxml
pandas